Fix issue with going back and organise code a bit.
When using the back button, the animations would not be completely correct. Now
they are. Also, the attributes of an element that can be used for navigation
have moved to a helper so that it is easier to create new navigation elements.
Finally, navigation is triggered not only by links in the header, but also by
links with the `nav-link` class. This will be used later (see next commit).