Fix navigation animation bug in Chrome.
Chrome will fire a `popstate` event even when navigating from some other
non-denvelop page to a denvelop page. This triggered an unwanted animation of
content. Now, no animation is done when navigating to the current page, which
fixes the problem. Note that navigating to the page itself normally does not
occur, as the site does not include self-links (hopefully at least...).