Themability!
It is now possible, at least on the Dutch page, to select a theme. Switching
happens with a short transition. What is left is adding this to the English
page, finishing the support for inverted images (for GitHub/Facebook/Twitter
logos mainly) and maybe creating a command that does the switching.