Many improvements to Desh navigation
Fix #3 and a couple of other small bugs. Also, Desh now has an internal
sitemap that is used for navigation instead of querying the page DOM.
This means that full site navigation is now possible, instead of only
the current language pages. Because of this change, the `cd-lang`
command has been rendered obsolete and is removed.