Fix spacing problems with floats
The theme swatches were overflowing out of their page because their
parent element did not have a clearfix. The same held true for the
`.floaty-columns` containers. The latter caused pages to run below the
footer. After fixing that, I noticed that there was extra space between
the content and footer when the footer was only one line. This is now
dynamic, so that the space between the content and footer is always the
same, not matter if there are 1 or 2 lines in the footer.