Improve/fix blog display
First of all, one of the pages with four images was not displaying
correctly anymore because of my changes and using a helper function.

Secondly, pages were not aligned correctly in Chrome, because
apparently the font rendering is slightly different in Firefox and
Chrome. Was to be expected of course. May have to test other browsers
too actually. Oh well.